v0.3.1: fix CI workflows, clippy lints, Docker build
- Fix 15 new clippy lints from Rust 1.95 (io_other_error,
  needless_range_loop, field_reassign_with_default, len_without_is_empty,
  collapsible_if, unnecessary_cast, manual_is_multiple_of)
- Add missing aarch64 target install in CI.yml
- Remove benches COPY from Dockerfile (conflicts with .dockerignore)
- Multi-arch Docker publish (linux/amd64 + linux/arm64) via release.yml
- Release job only runs on v* tags, not main pushes

+++ src/net/protocols/beast.rs
@@ -27,7 +27,7 @@ pub fn encode_beast_output(msg: &DecodedMessage) -> Vec<u8> {
     out.extend_from_slice(&now.to_be_bytes()[2..]); // last 6 bytes
 
     // Message type — derive from payload first byte (DF type)
-    let msg_type = msg.data.first().map(|b| (b >> 3) & 0x1F).unwrap_or(0) as u8;
+    let msg_type = msg.data.first().map(|b| (b >> 3) & 0x1F).unwrap_or(0);
     out.push(msg_type);
 
     // Payload
